#1 [refactor] application 설정 및 activity 이동
