+++ app/src/main/AndroidManifest.xml
@@ -2,7 +2,10 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools">
 
+    <uses-permission android:name="android.permission.INTERNET" />
+
     <application
+        android:name=".app.Application"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
@@ -13,7 +16,7 @@
         android:theme="@style/Theme.Softie"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name="com.sopetit.feature.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.Softie">
